niwa-api/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
websockets
google-genai
numpy
opencv-python
Pillow
exifread
ImageHash
